Add filters
when using the syntax <%= epression | filter | filter %> filters are
called in this order. a filter function has this prototype
function(string) where is string is the result of the expression. It
should return a string.
one can register a filter by calling blue.registerFilter(name, func)
where name is the name of the filter and func, the function to be called